+++ isthmus/src/main/java/io/substrait/isthmus/SubstraitToCalcite.java
@@ -1,6 +1,5 @@
 package io.substrait.isthmus;
 
-import io.substrait.extension.SimpleExtension;
 import io.substrait.isthmus.SubstraitRelNodeConverter.Context;
 import io.substrait.plan.Plan;
 import io.substrait.relation.Rel;
@@ -16,103 +15,33 @@ import org.apache.calcite.rel.type.RelDataType;
 import org.apache.calcite.rel.type.RelDataTypeFactory;
 import org.apache.calcite.rel.type.RelDataTypeField;
 import org.apache.calcite.sql.SqlKind;
-import org.apache.calcite.tools.Frameworks;
 import org.apache.calcite.tools.RelBuilder;
 import org.apache.calcite.util.Pair;
 
 /**
  * Converts between Substrait {@link Rel}s and Calcite {@link RelNode}s.
  *
- * <p>Can be extended to customize the {@link RelBuilder} and {@link SubstraitRelNodeConverter} used
- * in the conversion.
+ * <p>Conversion behaviours can be customized using a {@link ConverterProvider}
  */
 public class SubstraitToCalcite {
 
-  protected final SimpleExtension.ExtensionCollection extensions;
   protected final RelDataTypeFactory typeFactory;
-  protected final TypeConverter typeConverter;
   protected final Prepare.CatalogReader catalogReader;
-  protected final FeatureBoard featureBoard;
+  protected ConverterProvider converterProvider;
 
-  public SubstraitToCalcite(
-      SimpleExtension.ExtensionCollection extensions, RelDataTypeFactory typeFactory) {
-    this(extensions, typeFactory, TypeConverter.DEFAULT, null);
+  public SubstraitToCalcite(ConverterProvider converterProvider) {
+    this(converterProvider, null);
   }
 
   public SubstraitToCalcite(
-      SimpleExtension.ExtensionCollection extensions,
-      RelDataTypeFactory typeFactory,
-      Prepare.CatalogReader catalogReader) {
-    this(extensions, typeFactory, TypeConverter.DEFAULT, catalogReader);
-  }
-
-  public SubstraitToCalcite(
-      SimpleExtension.ExtensionCollection extensions,
-      RelDataTypeFactory typeFactory,
-      TypeConverter typeConverter) {
-    this(extensions, typeFactory, typeConverter, null);
-  }
-
-  public SubstraitToCalcite(
-      SimpleExtension.ExtensionCollection extensions,
-      RelDataTypeFactory typeFactory,
-      TypeConverter typeConverter,
-      Prepare.CatalogReader catalogReader) {
-    this(
-        extensions,
-        typeFactory,
-        typeConverter,
-        catalogReader,
-        ImmutableFeatureBoard.builder().build());
-  }
-
-  public SubstraitToCalcite(
-      SimpleExtension.ExtensionCollection extensions,
-      RelDataTypeFactory typeFactory,
-      TypeConverter typeConverter,
-      Prepare.CatalogReader catalogReader,
-      FeatureBoard featureBoard) {
-    this.extensions = extensions;
-    this.typeFactory = typeFactory;
-    this.typeConverter = typeConverter;
+      ConverterProvider converterProvider, Prepare.CatalogReader catalogReader) {
+    this.converterProvider = converterProvider;
+    this.typeFactory = converterProvider.getTypeFactory();
     this.catalogReader = catalogReader;
-    this.featureBoard = featureBoard;
-  }
-
-  /**
-   * Extracts a {@link CalciteSchema} from a {@link Rel}
-   *
-   * <p>Override this method to customize schema extraction.
-   */
-  protected CalciteSchema toSchema(Rel rel) {
-    SchemaCollector schemaCollector = new SchemaCollector(typeFactory, typeConverter);
-    return schemaCollector.toSchema(rel);
-  }
-
-  /**
-   * Creates a {@link RelBuilder} from the extracted {@link CalciteSchema}
-   *
-   * <p>Override this method to customize the {@link RelBuilder}.
-   */
-  protected RelBuilder createRelBuilder(CalciteSchema schema) {
-    return RelBuilder.create(Frameworks.newConfigBuilder().defaultSchema(schema.plus()).build());
-  }
-
-  /**
-   * Creates a {@link SubstraitRelNodeConverter} from the {@link RelBuilder}
-   *
-   * <p>Override this method to customize the {@link SubstraitRelNodeConverter}.
-   */
-  protected SubstraitRelNodeConverter createSubstraitRelNodeConverter(RelBuilder relBuilder) {
-    return new SubstraitRelNodeConverter(extensions, typeFactory, relBuilder, featureBoard);
   }
 
   /**
    * Converts a Substrait {@link Rel} to a Calcite {@link RelNode}
-   *
-   * <p>Generates a {@link CalciteSchema} based on the contents of the {@link Rel}, which will be
-   * used to construct a {@link RelBuilder} with the required schema information to build {@link
-   * RelNode}s, and a then a {@link SubstraitRelNodeConverter} to perform the actual conversion.
    *
    * @param rel {@link Rel} to convert
    * @return {@link RelNode}
@@ -120,12 +49,13 @@ public class SubstraitToCalcite {
   public RelNode convert(Rel rel) {
     RelBuilder relBuilder;
     if (catalogReader != null) {
-      relBuilder = createRelBuilder(catalogReader.getRootSchema());
+      relBuilder = converterProvider.getRelBuilder(catalogReader.getRootSchema());
     } else {
-      CalciteSchema rootSchema = toSchema(rel);
-      relBuilder = createRelBuilder(rootSchema);
+      CalciteSchema rootSchema = converterProvider.getSchemaResolver().apply(rel);
+      relBuilder = converterProvider.getRelBuilder(rootSchema);
     }
-    SubstraitRelNodeConverter converter = createSubstraitRelNodeConverter(relBuilder);
+    SubstraitRelNodeConverter converter =
+        converterProvider.getSubstraitRelNodeConverter(relBuilder);
 
     return rel.accept(converter, Context.newContext());
   }