+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 Release Notes
 ---
 
+## [0.80.0](https://github.com/substrait-io/substrait-java/compare/v0.79.0...v0.80.0) (2026-02-24)
+
+### ⚠ BREAKING CHANGES
+
+* removed FeatureBoard class
+* all constructors consuming a FeatureBoard have been
+removed
+* a number of methods/constructors have been replaced
+with ConverterProvider equivalents
+
+### Features
+
+* introduce ConverterProvider to control conversion behaviour ([#649](https://github.com/substrait-io/substrait-java/issues/649)) ([bb58d42](https://github.com/substrait-io/substrait-java/commit/bb58d4233abb21d87b7d89f6778e3fa5043fcb9b))
+
+## [0.79.0](https://github.com/substrait-io/substrait-java/compare/v0.78.0...v0.79.0) (2026-02-22)
+
+### Features
+
+* **isthmus:** add method to convert Plan into SQL strings using dialect ([#704](https://github.com/substrait-io/substrait-java/issues/704)) ([e75b621](https://github.com/substrait-io/substrait-java/commit/e75b621c8a73e032ba76964df70527b62ac7adff))
+* **spark:** upgrade scala compiler to 2.13.18 ([#705](https://github.com/substrait-io/substrait-java/issues/705)) ([becd45e](https://github.com/substrait-io/substrait-java/commit/becd45e080085733222d4cb2b0b6d5f0438fa07c))
+
 ## [0.78.0](https://github.com/substrait-io/substrait-java/compare/v0.77.0...v0.78.0) (2026-02-01)
 
 ### Features